[#] Refactor component's methods location
