+++ client/src/components/ProgramDialog.vue
@@ -29,7 +29,7 @@
             <span>Delete</span>
           </button>
         </div>
-        <button type="button" class="mdc-button mdc-dialog__footer__button mdc-dialog__footer__button--cancel"><i class="material-icons mdc-button__icon">clear</i><span>Cancel</span></button>
+        <button type="button" class="mdc-button mdc-dialog__footer__button mdc-dialog__footer__button--cancel" v-on:click="close"><i class="material-icons mdc-button__icon">clear</i><span>Cancel</span></button>
         <button type="button" class="mdc-button mdc-dialog__footer__button mdc-dialog__footer__button--accept mdc-button--raised" v-on:click="confirm"><i class="material-icons mdc-button__icon">check</i><span>Update</span></button>
       </footer>
     </div>
@@ -67,6 +67,7 @@ export default {
     },
     confirm: function () {
       Event.$emit('program.update', this.program, this.$el.querySelector('input[type=file]').files[0])
+      this.close()
     },
     fileChange: function (name, files) {
       if (files.length > 0) {
