+++ client/src/components/TopicDialog.vue
@@ -51,7 +51,7 @@
             <span>Delete</span>
           </button>
         </div>
-        <button type="button" class="mdc-button mdc-dialog__footer__button mdc-dialog__footer__button--cancel"><i class="material-icons mdc-button__icon">clear</i><span>Cancel</span></button>
+        <button type="button" class="mdc-button mdc-dialog__footer__button mdc-dialog__footer__button--cancel" v-on:click="close"><i class="material-icons mdc-button__icon">clear</i><span>Cancel</span></button>
         <button type="button" class="mdc-button mdc-dialog__footer__button mdc-dialog__footer__button--accept mdc-button--raised" v-on:click="confirm"><i class="material-icons mdc-button__icon">check</i><span>Update</span></button>
       </footer>
     </div>
@@ -93,6 +93,7 @@ export default {
     },
     confirm: function () {
       Event.$emit('topic.update', this.topic, this.medias)
+      this.close()
     },
     deleteTopic: function () {
       Event.$emit('topic.delete', this.topic)
